chore: refactors to use updateObject utility across the app
