+++ src/containers/Auth/Auth.js
@@ -6,6 +6,7 @@ import * as actions from "../../store/actions/index";
 import Button from "../../components/UI/Button/Button";
 import classes from "./Auth.css";
 import Input from "../../components/UI/Input/Input";
+import { updateObject } from "../../shared/utility";
 import Spinner from "../../components/UI/Spinner/Spinner";
 
 class Auth extends Component {
@@ -76,18 +77,16 @@ class Auth extends Component {
   }
 
   inputChangedHandler = (event, controlName) => {
-    const updatedControls = {
-      ...this.state.controls,
-      [controlName]: {
-        ...this.state.controls[controlName],
+    const updatedControls = updateObject(this.state.controls, {
+      [controlName]: updateObject(this.state.controls[controlName], {
         value: event.target.value,
         valid: this.checkValidity(
           event.target.value,
           this.state.controls[controlName].validation
         ),
         touched: true
-      }
-    };
+      })
+    });
     this.setState({ controls: updatedControls });
   };
